feat: render print_format_builder_beta via Chrome instead of WeasyPrint
Drop the WeasyPrint dependency for print_format_builder_beta and route it
through the existing Chromium-based PDF generator. WeasyPrint produced
inconsistent layout vs. the standard print path and could not share the
letterhead / page-number plumbing already in pdf_generator/browser.py.

- Add frappe/utils/print_format_generator.py with a Chrome-aware
  PrintFormatGenerator. The beta builder always uses overlay mode
  (#header-html / #footer-html wrappers) so headers/footers repeat on
  every page regardless of the global "Repeat Header and Footer"
  setting, and the page-number HTML is emitted at the top of the header
  so "Top Left/Center/Right" renders above the letterhead.
- Delete frappe/utils/weasyprint.py and drop the weasyprint dependency
  from pyproject.toml. Redirect callers in print_utils.py, printview.py,
  print.js, and the print_format doctype to the new generator.
- Guard the @page page-number CSS in templates/print_format/print_format.css
  with {% if not for_chrome %} so it only fires for WeasyPrint/browser
  print, not the Chrome body PDF (which has marginTop=0 and would render
  the counter inside the content area).
