+++ pyproject.toml
@@ -26,8 +26,6 @@ dependencies = [
     "PyQRCode~=1.2.1",
     "PyYAML~=6.0.3",
     "RestrictedPython~=8.1",
-    "WeasyPrint==68.0",
-    "pydyf==0.12.1",
     "Werkzeug==3.1.6",
     "Whoosh~=2.7.4",
     "beautifulsoup4~=4.13.5",
@@ -281,7 +279,6 @@ module = [
     "sqlparse",
     "terminaltables",
     "traceback_with_variables",
-    "weasyprint",
     "whoosh.fields",
     "whoosh.index",
     "whoosh.qparser",
